Small refactoring of API-mark
