--- map/styled_point.hpp
@@ -1,33 +1,0 @@
-#pragma once
-
-#include "map/user_mark.hpp"
-#include "map/user_mark_container.hpp"
-
-#include "search/result.hpp"
-
-#include "indexer/feature.hpp"
-
-#include "geometry/point2d.hpp"
-
-#include "std/string.hpp"
-#include "std/unique_ptr.hpp"
-#include "std/utility.hpp"
-
-namespace style
-{
-// Fixes icons which are not supported by MapsWithMe.
-string GetSupportedStyle(string const & s, string const & context, string const & fallback);
-// Default icon.
-string GetDefaultStyle();
-
-class StyledPoint : public UserMark
-{
-public:
-  StyledPoint(m2::PointD const & ptOrg, UserMarkContainer * container);
-
-  m2::PointD const & GetPixelOffset() const override;
-
-  /// @return name of icon, or empty string for plain circle.
-  virtual string const & GetStyle() const = 0;
-};
-}  // namespace style
