+++ map/user_mark_container.hpp
@@ -116,3 +116,12 @@ public:
 protected:
   UserMark * AllocateUserMark(m2::PointD const & ptOrg) override;
 };
+
+class ApiUserMarkContainer : public UserMarkContainer
+{
+public:
+  ApiUserMarkContainer(double layerDepth, Framework & framework);
+
+protected:
+  UserMark * AllocateUserMark(m2::PointD const & ptOrg) override;
+};
